+++ flybywire-aircraft-a320-neo/html_ui/Pages/VCockpit/Instruments/Airliners/FlyByWire_A320_Neo/CDU/A320_Neo_CDU_ProgressPage.js
@@ -3,10 +3,9 @@ class CDUProgressPage {
         mcdu.clearDisplay();
         mcdu.page.Current = mcdu.page.ProgressPage;
         mcdu.activeSystem = 'FMGC';
-        const flightPhase = "CRZ";
         const flightNo = SimVar.GetSimVarValue("ATC FLIGHT NUMBER", "string");
         const flMax = mcdu.getMaxFlCorrected();
-        const flOpt = (mcdu._zeroFuelWeightZFWCGEntered && mcdu._blockFuelEntered && (mcdu.isAllEngineOn() || Simplane.getIsGrounded())) ? "FL" + (Math.floor(flMax / 5) * 5).toString() + "[color]green" : "-----";
+        const flOpt = (mcdu._zeroFuelWeightZFWCGEntered && mcdu._blockFuelEntered && (mcdu.isAllEngineOn() || Simplane.getIsGrounded())) ? "{green}FL" + (Math.floor(flMax / 5) * 5).toString() + "{end}" : "-----";
         let flCrz = "-----";
         switch (mcdu.currentFlightPhase) {
             case FmgcFlightPhases.PREFLIGHT:
@@ -31,6 +30,32 @@ class CDUProgressPage {
                 break;
             }
         }
+        let flightPhase;
+        switch (mcdu.currentFlightPhase) {
+            case FmgcFlightPhases.PREFLIGHT:
+            case FmgcFlightPhases.TAKEOFF:
+                flightPhase = "TO";
+                break;
+            case FmgcFlightPhases.CLIMB:
+                flightPhase = "CLB";
+                break;
+            case FmgcFlightPhases.CRUISE:
+                flightPhase = "CRZ";
+                break;
+            case FmgcFlightPhases.DESCENT:
+                flightPhase = "DES";
+                break;
+            case FmgcFlightPhases.APPROACH:
+                flightPhase = "APPR";
+                break;
+            case FmgcFlightPhases.GOAROUND:
+                flightPhase = "GA";
+                break;
+            default:
+                flightPhase = "";
+                break;
+        }
+
         mcdu.onLeftInput[0] = (value) => {
             if (mcdu.trySetCruiseFlCheckInput(value)) {
                 CDUProgressPage.ShowPage(mcdu);
@@ -49,9 +74,9 @@ class CDUProgressPage {
             CDUProgressPage.ShowPredictiveGPSPage(mcdu);
         };
         mcdu.setTemplate([
-            ["{green}ECON " + flightPhase + "{end} " + flightNo],
-            ["\xa0" + flightPhase, "REC MAX\xa0", "OPT"],
-            [flCrz, "FL" + flMax.toString() + "\xa0[color]magenta", flOpt],
+            ["{green}" + flightPhase.padStart(15, "\xa0") + "{end}\xa0" + flightNo.padEnd(11, "\xa0")],
+            ["\xa0" + "CRZ\xa0", "OPT\xa0\xa0\xa0\xa0REC MAX"],
+            [flCrz, flOpt + "\xa0\xa0\xa0\xa0" + "{magenta}FL" + flMax.toString() + "\xa0{end}"],
             [""],
             ["<REPORT", ""],
             ["\xa0POSITION UPDATE AT"],